climate-intelligence-platform/requirements.txt
# Core AI and ML
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Web Framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
gradio>=4.0.0
websockets>=11.0.0

# Data Processing and Analytics
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Database and Storage
sqlalchemy>=2.0.0
redis>=5.0.0
psycopg2-binary>=2.9.0

# External APIs and Services
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
geopy>=2.4.0
python-dotenv>=1.0.0

# Climate and Environmental Data
meteostat>=1.6.0
pyowm>=3.3.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
markdown>=3.5.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0
sentry-sdk>=1.38.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Deployment
docker>=6.1.0
gunicorn>=21.2.0
celery>=5.3.0

# Additional utilities
python-multipart>=0.0.6
jinja2>=3.1.0
pydantic>=2.4.0
typer>=0.9.0
rich>=13.6.0